+++ lib/strings.dart
@@ -25,7 +25,7 @@ bool isEmpty(String s) => s == null || s.isEmpty;
 bool isNotEmpty(String s) => s != null && s.isNotEmpty;
 
 /// Returns a string with characters from the given [s] in reverse order.
-String flip(String s) {
+String reverse(String s) {
   if (s == null || s == '') return s;
   StringBuffer sb = new StringBuffer();
   var runes = s.runes;
@@ -35,11 +35,17 @@ String flip(String s) {
   return sb.toString();
 }
 
+/// Returns a string with characters from the given [s] in reverse order.
+///
+/// DEPRECATED: use [reverse] instead.
+@deprecated
+String flip(String s) => reverse(s);
+
 /// Concatenates [s] to itself a given number of [times]. Empty and null
 /// strings will always result in empty and null strings respectively no matter
 /// how many [times] they are [repeat]ed.
 ///
-/// If [times] is negative, returns the [flip]ped string repeated given number
+/// If [times] is negative, returns the reversed string repeated given number
 /// of [times].
 ///
 /// DEPRECATED: use the `*` operator on [String].
@@ -47,7 +53,7 @@ String flip(String s) {
 String repeat(String s, int times) {
   if (s == null || s == '') return s;
   if (times < 0) {
-    return repeat(flip(s), -times);
+    return repeat(reverse(s), -times);
   }
   StringBuffer sink = new StringBuffer();
   _repeat(sink, s, times);
@@ -79,7 +85,7 @@ String loop(String s, int from, [int to]) {
     throw new ArgumentError('Input string cannot be null or empty');
   }
   if (to != null && to < from) {
-    return loop(flip(s), -from, -to);
+    return loop(reverse(s), -from, -to);
   }
   int len = s.length;
   int leftFrag = from >= 0 ? from ~/ len : ((from - len) ~/ len);